+++ packages/database/src/databases/provider.ts
@@ -1670,12 +1670,20 @@ export class ProviderDatabase
 		});
 	}
 
-	/** @description Remove a detector key */
-	async removeDetectorKey(detectorKey: string): Promise<void> {
+	/**
+	 * @description Remove a detector key
+	 * @param detectorKey The detector key to remove
+	 * @param expirationInSeconds Optional expiration time in seconds (default is 10 minutes)
+	 * */
+	async removeDetectorKey(
+		detectorKey: string,
+		expirationInSeconds?: number,
+	): Promise<void> {
 		const filter: Pick<DetectorSchema, "detectorKey"> = { detectorKey };
 
-		// Instead of deleting, set the expiresAt field to expire in 10 minutes
-		const expiresAt = new Date(Date.now() + 10 * 60 * 1000); // 10 minutes from now
+		const expiresAt = new Date(
+			Date.now() + (expirationInSeconds || 10 * 60) * 1000,
+		);
 
 		await this.tables?.detector.updateOne(filter, {
 			$set: { expiresAt },