+++ dev/flux/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @prosopo/flux
 
+## 2.6.21
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [1a23649]
+  - @prosopo/common@3.1.11
+  - @prosopo/dotenv@3.0.16
+  - @prosopo/util@3.0.14
+  - @prosopo/config@3.1.11
+
+## 2.6.20
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/common@3.1.10
+  - @prosopo/dotenv@3.0.15
+  - @prosopo/util@3.0.13
+  - @prosopo/config@3.1.10
+
+## 2.6.19
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [809b984]
+  - @prosopo/common@3.1.9
+  - @prosopo/dotenv@3.0.14
+  - @prosopo/util@3.0.12
+  - @prosopo/config@3.1.9
+
 ## 2.6.18
 ### Patch Changes