+++ dev/lint/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @prosopo/lint
 
+## 2.8.6
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [1a23649]
+  - @prosopo/util@3.0.14
+  - @prosopo/config@3.1.11
+
+## 2.8.5
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/util@3.0.13
+  - @prosopo/config@3.1.10
+
+## 2.8.4
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- 6339d8e: Allow webp
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [809b984]
+  - @prosopo/util@3.0.12
+  - @prosopo/config@3.1.9
+
 ## 2.8.3
 ### Patch Changes