+++ dev/vite-plugin-watch-workspace/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prosopo/vite-plugin-watch-workspace
 
+## 2.6.17
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+
+## 2.6.16
+### Patch Changes
+
+- 657a827: Release 3.2.2
+
+## 2.6.15
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+
 ## 2.6.14
 ### Patch Changes