+++ packages/api/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @prosopo/api
 
+## 3.1.15
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types@3.1.4
+  - @prosopo/user-access-policy@3.5.10
+  - @prosopo/config@3.1.11
+
+## 3.1.14
+### Patch Changes
+
+- 36b23e0: Fix entropy. Fix api call. Persist ja4 through logs.
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/user-access-policy@3.5.9
+  - @prosopo/types@3.1.3
+  - @prosopo/config@3.1.10
+
+## 3.1.13
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/user-access-policy@3.5.8
+  - @prosopo/types@3.1.2
+  - @prosopo/config@3.1.9
+
 ## 3.1.12
 ### Patch Changes