+++ packages/file-server/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @prosopo/file-server
 
+## 2.6.16
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [1a23649]
+  - @prosopo/util@3.0.14
+  - @prosopo/config@3.1.11
+
+## 2.6.15
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/util@3.0.13
+  - @prosopo/config@3.1.10
+
+## 2.6.14
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [809b984]
+  - @prosopo/util@3.0.12
+  - @prosopo/config@3.1.9
+
 ## 2.6.13
 ### Patch Changes