+++ packages/procaptcha-bundle/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @prosopo/procaptcha-bundle
 
+## 3.0.25
+### Patch Changes
+
+- 0d1a33e: Adding ipcomparison service with user features
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/locale@3.1.11
+  - @prosopo/types@3.1.4
+  - @prosopo/procaptcha-frictionless@2.7.4
+  - @prosopo/procaptcha-common@2.8.4
+  - @prosopo/procaptcha-react@2.8.4
+  - @prosopo/widget-skeleton@2.6.12
+  - @prosopo/procaptcha-pow@2.8.4
+  - @prosopo/dotenv@3.0.16
+  - @prosopo/util@3.0.14
+  - @prosopo/config@3.1.11
+
+## 3.0.24
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/procaptcha-frictionless@2.7.3
+  - @prosopo/procaptcha-common@2.8.3
+  - @prosopo/procaptcha-react@2.8.3
+  - @prosopo/widget-skeleton@2.6.11
+  - @prosopo/procaptcha-pow@2.8.3
+  - @prosopo/dotenv@3.0.15
+  - @prosopo/locale@3.1.10
+  - @prosopo/types@3.1.3
+  - @prosopo/util@3.0.13
+  - @prosopo/config@3.1.10
+
+## 3.0.23
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/procaptcha-frictionless@2.7.2
+  - @prosopo/procaptcha-common@2.8.2
+  - @prosopo/procaptcha-react@2.8.2
+  - @prosopo/widget-skeleton@2.6.10
+  - @prosopo/procaptcha-pow@2.8.2
+  - @prosopo/dotenv@3.0.14
+  - @prosopo/locale@3.1.9
+  - @prosopo/types@3.1.2
+  - @prosopo/util@3.0.12
+  - @prosopo/config@3.1.9
+
 ## 3.0.22
 ### Patch Changes