+++ packages/procaptcha-common/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @prosopo/procaptcha-common
 
+## 2.8.4
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types@3.1.4
+  - @prosopo/widget-skeleton@2.6.12
+  - @prosopo/load-balancer@2.7.4
+  - @prosopo/account@2.7.21
+  - @prosopo/config@3.1.11
+
+## 2.8.3
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/widget-skeleton@2.6.11
+  - @prosopo/load-balancer@2.7.3
+  - @prosopo/account@2.7.20
+  - @prosopo/types@3.1.3
+  - @prosopo/config@3.1.10
+
+## 2.8.2
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/widget-skeleton@2.6.10
+  - @prosopo/load-balancer@2.7.2
+  - @prosopo/account@2.7.19
+  - @prosopo/types@3.1.2
+  - @prosopo/config@3.1.9
+
 ## 2.8.1
 ### Patch Changes