+++ packages/procaptcha/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @prosopo/procaptcha
 
+## 2.8.4
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types@3.1.4
+  - @prosopo/procaptcha-common@2.8.4
+  - @prosopo/util-crypto@13.5.13
+  - @prosopo/datasets@3.0.25
+  - @prosopo/common@3.1.11
+  - @prosopo/util@3.0.14
+  - @prosopo/api@3.1.15
+  - @prosopo/config@3.1.11
+
+## 2.8.3
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [36b23e0]
+- Updated dependencies [657a827]
+  - @prosopo/api@3.1.14
+  - @prosopo/procaptcha-common@2.8.3
+  - @prosopo/util-crypto@13.5.12
+  - @prosopo/datasets@3.0.24
+  - @prosopo/common@3.1.10
+  - @prosopo/types@3.1.3
+  - @prosopo/util@3.0.13
+  - @prosopo/config@3.1.10
+
+## 2.8.2
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/procaptcha-common@2.8.2
+  - @prosopo/util-crypto@13.5.11
+  - @prosopo/datasets@3.0.23
+  - @prosopo/common@3.1.9
+  - @prosopo/types@3.1.2
+  - @prosopo/util@3.0.12
+  - @prosopo/api@3.1.13
+  - @prosopo/config@3.1.9
+
 ## 2.8.1
 ### Patch Changes