+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,88 @@
 # @prosopo/provider
 
+## 3.8.4
+### Patch Changes
+
+- 0d1a33e: Adding ipcomparison service with user features
+- 0d1a33e: Adding ip comparison service
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types-database@3.1.5
+  - @prosopo/locale@3.1.11
+  - @prosopo/types@3.1.4
+  - @prosopo/api-express-router@3.0.16
+  - @prosopo/user-access-policy@3.5.10
+  - @prosopo/load-balancer@2.7.4
+  - @prosopo/util-crypto@13.5.13
+  - @prosopo/api-route@2.6.19
+  - @prosopo/types-env@2.7.29
+  - @prosopo/database@3.2.4
+  - @prosopo/datasets@3.0.25
+  - @prosopo/keyring@2.8.18
+  - @prosopo/common@3.1.11
+  - @prosopo/util@3.0.14
+  - @prosopo/api@3.1.15
+  - @prosopo/env@3.2.4
+  - @prosopo/config@3.1.11
+
+## 3.8.3
+### Patch Changes
+
+- 36b23e0: Fix entropy. Fix api call. Persist ja4 through logs.
+- 657a827: Release 3.2.2
+- 4aac849: Do not error when IPs don't match
+- Updated dependencies [36b23e0]
+- Updated dependencies [a8a9251]
+- Updated dependencies [657a827]
+  - @prosopo/api@3.1.14
+  - @prosopo/types-database@3.1.4
+  - @prosopo/api-express-router@3.0.15
+  - @prosopo/user-access-policy@3.5.9
+  - @prosopo/load-balancer@2.7.3
+  - @prosopo/util-crypto@13.5.12
+  - @prosopo/api-route@2.6.18
+  - @prosopo/types-env@2.7.28
+  - @prosopo/database@3.2.3
+  - @prosopo/datasets@3.0.24
+  - @prosopo/keyring@2.8.17
+  - @prosopo/common@3.1.10
+  - @prosopo/locale@3.1.10
+  - @prosopo/types@3.1.3
+  - @prosopo/util@3.0.13
+  - @prosopo/env@3.2.3
+  - @prosopo/config@3.1.10
+
+## 3.8.2
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- 1249ce0: Be more lenient with random provider selection
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/api-express-router@3.0.14
+  - @prosopo/user-access-policy@3.5.8
+  - @prosopo/types-database@3.1.3
+  - @prosopo/load-balancer@2.7.2
+  - @prosopo/util-crypto@13.5.11
+  - @prosopo/api-route@2.6.17
+  - @prosopo/types-env@2.7.27
+  - @prosopo/database@3.2.2
+  - @prosopo/datasets@3.0.23
+  - @prosopo/keyring@2.8.16
+  - @prosopo/common@3.1.9
+  - @prosopo/locale@3.1.9
+  - @prosopo/types@3.1.2
+  - @prosopo/util@3.0.12
+  - @prosopo/api@3.1.13
+  - @prosopo/env@3.2.2
+  - @prosopo/config@3.1.9
+
 ## 3.8.1
 ### Patch Changes