+++ packages/types-database/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @prosopo/types-database
 
+## 3.1.5
+### Patch Changes
+
+- 0d1a33e: Adding ipcomparison service with user features
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/locale@3.1.11
+  - @prosopo/types@3.1.4
+  - @prosopo/user-access-policy@3.5.10
+  - @prosopo/common@3.1.11
+  - @prosopo/config@3.1.11
+
+## 3.1.4
+### Patch Changes
+
+- a8a9251: Add index on IP parts
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/user-access-policy@3.5.9
+  - @prosopo/common@3.1.10
+  - @prosopo/locale@3.1.10
+  - @prosopo/types@3.1.3
+  - @prosopo/config@3.1.10
+
+## 3.1.3
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- 1249ce0: Be more lenient with random provider selection
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/user-access-policy@3.5.8
+  - @prosopo/common@3.1.9
+  - @prosopo/locale@3.1.9
+  - @prosopo/types@3.1.2
+  - @prosopo/config@3.1.9
+
 ## 3.1.2
 ### Patch Changes