+++ packages/types-env/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @prosopo/types-env
 
+## 2.7.29
+### Patch Changes
+
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/types-database@3.1.5
+  - @prosopo/types@3.1.4
+  - @prosopo/keyring@2.8.18
+  - @prosopo/common@3.1.11
+  - @prosopo/config@3.1.11
+
+## 2.7.28
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [a8a9251]
+- Updated dependencies [657a827]
+  - @prosopo/types-database@3.1.4
+  - @prosopo/keyring@2.8.17
+  - @prosopo/common@3.1.10
+  - @prosopo/types@3.1.3
+  - @prosopo/config@3.1.10
+
+## 2.7.27
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [1249ce0]
+- Updated dependencies [809b984]
+  - @prosopo/types-database@3.1.3
+  - @prosopo/keyring@2.8.16
+  - @prosopo/common@3.1.9
+  - @prosopo/types@3.1.2
+  - @prosopo/config@3.1.9
+
 ## 2.7.26
 ### Patch Changes