+++ packages/types/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @prosopo/types
 
+## 3.1.4
+### Patch Changes
+
+- 0d1a33e: Adding ipcomparison service with user features
+- 0d1a33e: Adding ip comparison service
+- 1a23649: Release 3.2.3
+- Updated dependencies [0d1a33e]
+- Updated dependencies [1a23649]
+  - @prosopo/locale@3.1.11
+  - @prosopo/util-crypto@13.5.13
+  - @prosopo/util@3.0.14
+  - @prosopo/config@3.1.11
+
+## 3.1.3
+### Patch Changes
+
+- 657a827: Release 3.2.2
+- Updated dependencies [657a827]
+  - @prosopo/util-crypto@13.5.12
+  - @prosopo/locale@3.1.10
+  - @prosopo/util@3.0.13
+  - @prosopo/config@3.1.10
+
+## 3.1.2
+### Patch Changes
+
+- 4440947: fix type-only tsc compilation
+- 7bdaca6: Release 3.2.1
+- 1249ce0: Be more lenient with random provider selection
+- Updated dependencies [4440947]
+- Updated dependencies [7bdaca6]
+- Updated dependencies [809b984]
+- Updated dependencies [809b984]
+  - @prosopo/util-crypto@13.5.11
+  - @prosopo/locale@3.1.9
+  - @prosopo/util@3.0.12
+  - @prosopo/config@3.1.9
+
 ## 3.1.1
 ### Patch Changes